Déplace la route "/motDePasse/edition" vers le routeur dédié
